feat: new spaceview and keypoint mock data
