+++ ZalZaZa/ZalZaZa/Data/keypoints.json
@@ -1,0 +1,43 @@
+/* 
+  keypoints.json
+  ZalZaZa
+
+  Created by Sulim Chun on 2022/11/14.
+  
+*/
+
+{[
+    "0": {"x": 408.7763864712414, "y": 309.9413242968486, "z": -988000.3529131957, "score": 0.9999776266520978, "name": "nose"},
+    "1": {"x": 439.7090541894572, "y": 283.81495369387596, "z": -982622.6608739386, "score": 0.9999746560907121, "name": "left_eye_inner"},
+    "2": {"x": 456.3262622477794, "y": 283.03441333091047, "z": -982439.3048867706, "score": 0.9999577941432881, "name": "left_eye"},
+    "3": {"x": 470.2569943726135, "y": 282.0026389310943, "z": -982474.5934787771, "score": 0.9999522622538326, "name": "left_eye_outer"},
+    "4": {"x": 386.0352446340851, "y": 277.50789357624575, "z": -984942.3030141006, "score": 0.9999822142918359, "name": "right_eye_inner"},
+    "5": {"x": 371.49575280936693, "y": 273.37441481575297, "z": -985077.9108836108, "score": 0.9999743864118458, "name": "right_eye"},
+    "6": {"x": 360.91367056849407, "y": 267.0640140376528, "z": -984934.2646104546, "score": 0.9999726855693316, "name": "right_eye_outer"},
+    "7": {"x": 485.43580348579667, "y": 282.9636959691165, "z": -829663.575557519, "score": 0.9999507484257575, "name": "left_ear"},
+    "8": {"x": 333.6387785508296, "y": 252.114319205181, "z": -840808.6614447972, "score": 0.9999755271753998, "name": "right_ear"},
+    "9": {"x": 425.24388146993317, "y": 353.62794827874325, "z": -920044.6530972994, "score": 0.9995684370024343, "name": "mouth_left"},
+    "10": {"x": 373.62877823846844, "y": 341.63026212472846, "z": -923107.7671906913, "score": 0.9995724657827795, "name": "mouth_right"},
+    "11": {"x": 508.87057894515755, "y": 387.18142558410057, "z": -623104.2941526884, "score": 0.9964150027036935, "name": "left_shoulder"},
+    "12": {"x": 249.79330665967439, "y": 327.1236645617291, "z": -638011.915634608, "score": 0.993908780560876, "name": "right_shoulder"},
+    "13": {"x": 593.4946610638974, "y": 490.1777202109482, "z": -561107.1414235181, "score": 0.8815400323470103, "name": "left_elbow"},
+    "14": {"x": 129.17187349115267, "y": 392.7312885396444, "z": -539369.0472113774, "score": 0.8955882409004109, "name": "right_elbow"},
+    "15": {"x": 546.6788347310143, "y": 522.3086392461735, "z": -570625.7367170395, "score": 0.32681351986148244, "name": "left_wrist"},
+    "16": {"x": 132.60258508445304, "y": 403.83400363770335, "z": -476515.7241098584, "score": 0.3611521782608685, "name": "right_wrist"},
+    "17": {"x": 537.6045953910364, "y": 533.1476669759451, "z": -578033.6079811618, "score": 0.19976155399481152, "name": "left_pinky"},
+    "18": {"x": 130.21885485735254, "y": 401.52694760729844, "z": -480440.7962262244, "score": 0.214466862807921, "name": "right_pinky"},
+    "19": {"x": 522.1814986972532, "y": 516.3685924466618, "z": -570919.7011383784, "score": 0.20302441502725108, "name": "left_index"},
+    "20": {"x": 147.60684943413196, "y": 393.9872176376921, "z": -476983.15728188056, "score": 0.21260708534385056, "name": "right_index"},
+    "21": {"x": 518.677980019865, "y": 513.5708351079284, "z": -566113.258254233, "score": 0.1982804196398317, "name": "left_thumb"},
+    "22": {"x": 153.51650389332863, "y": 398.2947285902949, "z": -473358.6812698614, "score": 0.21334868403803564, "name": "right_thumb"},
+    "23": {"x": 387.3375128291258, "y": 568.223548561019, "z": 5925.295727643588, "score": 0.6240367917375235, "name": "left_hip"},
+    "24": {"x": 255.68499384382193, "y": 543.2593041942948, "z": -5255.370771773605, "score": 0.6721891593136077, "name": "right_hip"},
+    "25": {"x": 362.19507503799093, "y": 691.7715835469733, "z": 350877.2033777215, "score": 0.5945864373527934, "name": "left_knee"},
+    "26": {"x": 228.10705820842827, "y": 696.1161962898966, "z": 331723.07411363645, "score": 0.5027394616664259, "name": "right_knee"},
+    "27": {"x": 328.53795508550263, "y": 788.7745727472873, "z": 786338.6698898575, "score": 0.10743505314706168, "name": "left_ankle"},
+    "28": {"x": 226.79815956294647, "y": 791.4976350700567, "z": 714352.2733336851, "score": 0.0851652254500297, "name": "right_ankle"},
+    "29": {"x": 314.90415987555025, "y": 801.9943659969907, "z": 825739.4680495904, "score": 0.023348833965254664, "name": "left_heel"},
+    "30": {"x": 227.17845344356107, "y": 801.4969676550304, "z": 749540.7068307396, "score": 0.03827528284497771, "name": "right_heel"},
+    "31": {"x": 307.955581342128, "y": 864.2292627258665, "z": 735030.0217208656, "score": 0.11363058575705115, "name": "left_foot_index"},
+    "32": {"x": 241.13929843040864, "y": 883.5799128348281, "z": 626634.0777697562, "score": 0.11446375615934483, "name": "right_foot_index"}
+]}
