+++ ZalZaZa/ZalZaZa/Data/keypoints.strings
@@ -1,0 +1,7 @@
+/* 
+  keypoints.strings
+  ZalZaZa
+
+  Created by Sulim Chun on 2022/11/14.
+  
+*/
